refactor: 상태 초기화를 위해 메인컨테이너를 CardRegisterPage로 분리
